requirements.txt
connexion==2.2.0
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
gym==0.10.5
numpy==1.14.5
pandas==0.23.3
Keras==2.2.4
python-chess==0.25.0
tensorboard==1.12.2
#torch==1.0.1
Werkzeug==0.14.1
